ttn2arduinocloud/requirements.txt
arduino_iot_cloud==1.0.0
paho_mqtt==1.6.1
pandas==2.2.2